Group by search

+++ backend/dsvendas/src/main/java/com/devsuperior/dsvendas/dto/SaleDTO.java
@@ -6,8 +6,8 @@ import com.devsuperior.dsvendas.entities.Sale;
 
 public class SaleDTO {    
     private Long id;
-    private Integer visited;
-    private Integer deals;
+    private Long visited;
+    private Long deals;
     private Double amount;
     private LocalDate date;
     private SellerDTO seller;
@@ -21,7 +21,7 @@ public class SaleDTO {
         this.seller = new SellerDTO(sale.getSeller());
     }
 
-    public SaleDTO(Long id, Integer visited, Integer deals, Double amount, LocalDate date, SellerDTO seller) {
+    public SaleDTO(Long id, Long visited, Long deals, Double amount, LocalDate date, SellerDTO seller) {
         this.id = id;
         this.visited = visited;
         this.deals = deals;
@@ -42,19 +42,19 @@ public class SaleDTO {
         this.id = id;
     }
 
-    public Integer getVisited() {
+    public Long getVisited() {
         return this.visited;
     }
 
-    public void setVisited(Integer visited) {
+    public void setVisited(Long visited) {
         this.visited = visited;
     }
 
-    public Integer getDeals() {
+    public Long getDeals() {
         return this.deals;
     }
 
-    public void setDeals(Integer deals) {
+    public void setDeals(Long deals) {
         this.deals = deals;
     }
